refactors fastai engine
